+++ com.ibm.streamsx.topology/opt/python/packages/streamsx/topology/topology.py
@@ -1782,6 +1782,80 @@ class Stream(_placement._Placement, object):
         oport = op.addOutputPort(name=_name, schema=schema)
         return Stream(self.topology, oport)._make_placeable()._layout('FlatMap', name=op.runtime_id, orig_name=name)._add_hints(hints)
     
+    def catch_exceptions(self, exception_type:str='streams', tuple_trace:bool=False, stack_trace:bool=False):
+        """ When applied to a primitive operator, exceptions of the specified type that are thrown by the operator while processing a tuple are caught. 
+
+        .. note:: You cannot use this on an operator without input streams.
+
+        Example using default values (tuple trace and stack trace disabled) and catch exceptions thrown by the Python primitive operator calling the ``map()`` transformation. This **map** callable raises a ValueError (*"invalid literal for int() with base 10: 'five'"*) when processing the sixt tuple. With ``catch_exceptions()`` applied the application is able to process all 10 tuples and does not stop processing.::
+
+           from typing import NamedTuple
+           class NumbersSchema(NamedTuple):
+              num: int
+
+           topo = Topology()
+           str_stream = topo.source(['0','1','2','3','4','five','6','7','8','9']).as_string()
+
+           num_stream = str_stream.map(lambda t: {'num': int(t)}, schema=NumbersSchema)
+           num_stream.catch_exceptions()
+
+           num_stream.print()
+
+        Example using the SPL operator Functor and enabled tuple trace::
+
+           topo = Topology()
+           str_stream = topo.source(['0','1','2','3','4','five','6','7','8','9']).as_string()
+
+           f = op.Map('spl.relational::Functor', s, schema='tuple<int64 num>')
+           f.num = f.output('(int64) string')
+           num_stream = f.stream
+           num_stream.catch_exceptions(tuple_trace=True)
+
+           num_stream.print()
+
+        Args:
+            exception_type(str): Indicates the type of exceptions to be caught by the run time environment. Supported options include:
+
+                * ``none``: No exceptions of any type are caught.
+
+                * ``streams``: Only IBM® Streams exceptions are caught. This includes exceptions that are thrown by SPL native functions from the standard toolkit, other exceptions that extend from the C++ SPL::SPLRuntimeException, and exceptions that extend from the Java com.ibm.streams.operator.DataException (extending from java.lang.RuntimeException).
+
+                * ``std``: Both IBM Streams and standard exceptions are caught. For C++, standard exception means std::exception. For Java, standard exception means all checked exceptions that inherit from java.lang.Exception.
+
+                * ``all``: In C++, any thrown exception is caught. For Java, any checked and unchecked exception that inherits from java.lang.Exception is caught.
+
+            tuple_trace(bool): Enables or disables the tracing of tuple data. Tracing of data can be enabled when tuples do not contain sensitive data and the data can show up into PE logs. Tuples are logged to the trace facility with the ERROR trace level.
+            stack_trace(bool): Enables or disables the printout of the stack trace to the Streams trace facility. Stack traces are printed to the Streams trace facility with the trace level ERROR.
+
+        Returns:
+            Stream: Returns this stream.
+
+        .. versionadded:: 2.1
+        """
+        if exception_type is None:
+           raise ValueError("Parameter exception_type must not be None.")
+        else:
+           if exception_type not in {'none', 'all', 'streams', 'std'}:
+              raise ValueError("Invalid value for parameter exception_type (supported options: 'none', 'all', 'streams', 'std').")
+
+        props = {'exception':exception_type}
+
+        if tuple_trace is not None:
+           if tuple_trace:
+              props['tupleTrace'] = 'true'
+           else:
+              props['tupleTrace'] = 'false'
+        if stack_trace is not None:
+           if not stack_trace:
+              props['stackTrace'] = 'false'
+           else:
+              props['stackTrace'] = 'true'
+
+        # create annotation dict
+        annotation = {'type':'catch', 'properties': props}
+        self.oport.operator._annotation(annotation)
+        return self._make_placeable()
+
     def isolate(self):
         """
         Guarantees that the upstream operation will run in a separate processing element from the downstream operation