+++ java/src/com/ibm/streamsx/topology/generator/spl/OperatorGenerator.java
@@ -114,6 +114,9 @@ class OperatorGenerator {
             else if (jstring(annotation, OpProperties.ANNOTATION_TYPE).equals("service")) {
             	appendServiceAnnotation(sb, properties);
             }
+            else if (jstring(annotation, OpProperties.ANNOTATION_TYPE).equals("catch")) {
+            	appendCatchAnnotation(sb, properties);
+            }
             else if (jstring(annotation, OpProperties.ANNOTATION_TYPE).equals("eventTime")) {
             	appendEventTimeAnnotation(sb, properties);
             }
@@ -256,6 +259,31 @@ class OperatorGenerator {
                 firstAttr = false;
             }
             sb.append("}\"");
+            first = false;
+    	}
+    }
+
+    private static void appendCatchAnnotation(StringBuilder sb, JsonObject properties) {
+    	boolean first = true;
+    	if (properties.has("exception")) {
+            sb.append("exception=");
+            sb.append(GsonUtilities.jstring(properties, "exception"));
+            first = false;
+    	}
+    	if (properties.has("tupleTrace")) {
+            if (!first) {
+                sb.append(',');
+            }
+            sb.append("tupleTrace=");
+            sb.append(GsonUtilities.jstring(properties, "tupleTrace"));
+            first = false;
+    	}
+    	if (properties.has("stackTrace")) {
+            if (!first) {
+                sb.append(',');
+            }
+            sb.append("stackTrace=");
+            sb.append(GsonUtilities.jstring(properties, "stackTrace"));
             first = false;
     	}
     }